Add 7.5.0 and linux support

+++ recipe/build.sh
@@ -19,28 +19,33 @@ cd build_conda
     --with-mpc=${PREFIX} \
     --with-isl=${PREFIX}
 
-# using || to quiet logs unless there is an issue
-{
-    make -j"${CPU_COUNT}" >& make_logs.txt
-} || {
-    tail -n 5000 make_logs.txt
-    exit 1
-}
+if [[ "$target_platform" == "osx-64" ]]; then
+  # using || to quiet logs unless there is an issue
+  {
+      make -j"${CPU_COUNT}" >& make_logs.txt
+  } || {
+      tail -n 5000 make_logs.txt
+      exit 1
+  }
 
-# using || to quiet logs unless there is an issue
-{
-    make install-strip >& make_install_logs.txt
-} || {
-    tail -n 5000 make_install_logs.txt
-    exit 1
-}
+  # using || to quiet logs unless there is an issue
+  {
+      make install-strip >& make_install_logs.txt
+  } || {
+      tail -n 5000 make_install_logs.txt
+      exit 1
+  }
+  rm $PREFIX/lib/libgomp.dylib
+  rm $PREFIX/lib/libgomp.1.dylib
+  ln -s $PREFIX/lib/libomp.dylib $PREFIX/lib/libgomp.dylib
+  ln -s $PREFIX/lib/libomp.dylib $PREFIX/lib/libgomp.1.dylib
 
-rm $PREFIX/lib/libgomp.dylib
-rm $PREFIX/lib/libgomp.1.dylib
-ln -s $PREFIX/lib/libomp.dylib $PREFIX/lib/libgomp.dylib
-ln -s $PREFIX/lib/libomp.dylib $PREFIX/lib/libgomp.1.dylib
+  pushd ${PREFIX}/lib
+    sed -i.bak "s@^\*lib.*@& -rpath $PREFIX/lib@" libgfortran.spec
+    rm libgfortran.spec.bak
+  popd
+else
+  make all-gcc -j${CPU_COUNT}
+  make install-gcc -j${CPU_COUNT}
+fi
 
-pushd ${PREFIX}/lib
-sed -i.bak "s@^\*lib.*@& -rpath $PREFIX/lib@" libgfortran.spec
-rm libgfortran.spec.bak
-popd
